+++ src/isotools/__init__.py
@@ -7,9 +7,10 @@ __version__ = distribution('isotools').version
 from .gene import Gene
 from .transcriptome import Transcriptome
 from .splice_graph import SegmentGraph, SegGraphNode
+from ._transcriptome_stats import estimate_tpm_threshold
 
 from ._transcriptome_filter import DEFAULT_GENE_FILTER, DEFAULT_TRANSCRIPT_FILTER, DEFAULT_REF_TRANSCRIPT_FILTER, ANNOTATION_VOCABULARY
 
 
-__all__ = ['Transcriptome', 'Gene', 'SegmentGraph', 'SegGraphNode',
+__all__ = ['Transcriptome', 'Gene', 'SegmentGraph', 'SegGraphNode', 'estimate_tpm_threshold',
            'DEFAULT_GENE_FILTER', 'DEFAULT_TRANSCRIPT_FILTER', 'DEFAULT_REF_TRANSCRIPT_FILTER', 'ANNOTATION_VOCABULARY']